class for running archiving (and cleanup) in batch
tarsnap::batch is a class (not a define) to run all archiving jobs in a
batch rather than randomized through multiple instances of tarsnap::periodic

We "simply" pass an Hash with Arrays of directories. Each key => [Array]
of the Hash will be processed serially.

we also clean up the our templates: sed can be used on both platforms!